Adjust project name to gstvimbasrc
The name of the resulting file has to match the name of the GStreamer
plugin in order for the plugin to be picked up correctly, because
GStreamer determines some symlinks directly from the filename [1]. The
simplest way to quickly do this was to adjust the project name for now.
This might change later on but that would mean that another way has to
be found to change the output binary file.

[1] https://github.com/GStreamer/gstreamer/commit/e7ede5a487d8c1571a4387a5f8eeb4d5e73a9553