Begin with Vimba integration
This only opens and closes the Vimba API and is mainly used as a test to
see that linking against the VimbaC library works as expected.

TODO: The CMake integration needs to be overhauled in order to also
support Linux and should generally not be as hardcoded as it is right
now. It would probably make sense to write a `FindVimba.cmake` file and
include this in the cmake subdirectory.