Implement opening of Camera
This is a first attempt at opening the camera connection. Opening was
moved from originally planned place in `vimbarsrc_init` because at that
point the `camera` property containing the required device ID has not
been set. It possibly even makes more sense to move the opening into the
`vimbasrc_start` function and close the camera connection in the
corresponding `vimbasrc_close`. Question is how this would influence
the capabilities I plan on polling from the camera for things like pixel
format.