Add basic `get_caps` functionality
This implementation does not truly represent the capabilites of the
connected camera but is only meant as a first foothold to get some
frames displayed. The values shown here work with an Alvium U-500m, but
adjusting the values for `width` and `height` should be enough to get
this working for other cameras as well.

Commandline used to get the first frames displayed on the screen (tested
on Windows):
```
gst-launch-1.0 vimbasrc camera=DEV_1AB22D01BBB8 ! video/x-raw,width=2592,height=1944 ! videoconvert ! queue ! autovideosink
```

If width and height are not explicitely set in the pipeline, the later
elements attempt to set their values to 1. As `set_caps` is not zet
implemented the recorded image data still contains the full sensor
resolution and the images are not displayed correctly.