Add small todo comment in `set_property`
If the camera cannot be openend it would be nice if the pipeline
immediatly halted/stopped. Currently the pipeline will continue with its
initialization and attempt to announce buffers in `vimbasrc_start`,
which fails and triggers the pipeline to halt. As the source cannot
really be used without a properly connected camera the markes position
would be a good place to halt the pipeline if possible.