Add control of ExposureAuto feature as property
This should be a good starting point for other enum features. The added
enum in `vimbasrc.h` should be reusable for all (or most) of the "auto"
features, as they all support the three modes listed. The actual string
conversion is defined in the `GEnumValue` in line 90.

One limitation of this approach is that this will probalby not be able
to support enum features where the available entries can only be
determined at runtime. For these features it would likely be easier to
define them as `string` properties and ensure that the passed string is
one of the valid entries.