Fix illegal variable declaration in switch cases
This behaviour is not allowed as the label of a switch case has to be
followed by a statement. A declaration is not legal and gcc refuses to
compile this. Adding a semicolon introduces an empty statement which in
turn makes the switch cases work as expected. Though it might make sense
to investigate this further to see if it can be restructured to avoid
the variable declaration.