Implement reporting of supported GStreamer formats
This gets the pixel formats supported by the camera and maps them to
compatible GStreamer formats. These compatible formats are then reported
as caps so that negotiation between elements works as inteded.

Setting of the pixelformat based on the negotiated format value needs to
be implemented in `gst_vimbasrc_set_caps`.