Reallocate new buffers if PayloadSize changed
If the PayloadSize increases after changing the pixel format or
width/height features of the camera, the existing buffers are not large
enough to accept the incoming frame. They are freed and new buffers are
allocated in their place. If the PayloadSize decreases, the buffers are
too large for the data which is not a problem. Therefore they are not
reallocated to save time on the `free` and `malloc` calls.