Add helper struct to vimbasrc for camera features
This struct will be a data cache to which new desired values will be
written via `gst_vimbasrc_set_property`. The settings are only applied
when the newly added `apply_feature_settings` function is called. This
buffering has the benefit, that it can more easily be adapted to e.g.
checks for an open camera connection.

The code should behave more or less the same as before, ROI settings may
still not work as expected since the order of offsetx/y and width/height
is not yet ideal.