Move camera open to `gst_vimbasrc_start`
This also fixes the previously mentioned problem of unexpected contents
of `vimbasrc->camera.id`. The problem was, that the used `get_string`
function returns a reference to the string inside the GValue object. If
the GValue object goes out  of scope, that reference becomes invalid.
Using explicit copies of the string with `dup_string` requires manual
freeing of the memory but fixes the problem of invalid memory access.