Again adjust log levels of some messages
Turns out not even ERROR level messages are displayed if the `GST_DEBUG`
environment variable is unset. So only critical errors that actually
will lead to inability to stream images are now left as ERROR level
messages. Other problems which will result in undesirable behaviour, but
not critical errors are changed to WARNING messages.