Implement Trigger features as element properties
This exposes the `TriggerActivation`, `TriggerMode`, `TriggerSelector`,
and `TriggerSource` camera feature as GStreamer element properties.
These are all enum features, which made it necessary to expose all
possible enum values for these features. The currenttly implemented
feature values are taken from the SFNC 2.4 and may not match our needs
perfectly. The number of "countable" trigger sources (e.g. Line0 -
Line3) may not count far enough and p ossibly this includes trigger
sources that are not implemented in any of our cameras but are available
as part of the SFNC.

This still requires some more helpful error messages that should ideally
display actual allowed values for the attempted feature setting.
Additionally the README should probably contain some more information on
how triggering can be used with the plugin.