Implement abort of `_create` on stream end
If the pipeline signals to the gst-vimbasrc element, that it should
change its state to PAUSED (or anything else besides PLAYING), no more
data should be pushed into the pipeline. This change leads to the
`_create` function aborting its wait for new frames when the state is no
longer PLAYING. This is espacially important for example if images are
hardware triggered because in that case the trigger generator might not
trigger a last image to actually finish the `_create` call what is
blocking the pipeline shutdown.