Add check if PixelFormat is available
In some cases cameras report capabilities for PixelFormats, that are not
currently settable. By adding a check if an enum entry we got from the
camera is currently available for setting, only pixelformats that can
actually be used now are added to the format mapping from which the src
pad caps are determined.