Update TriggerSources and TriggerSelectors
These are the values that are likely to be enough to cover Alvium
cameras and will likely also work for a large number of other cameras.
The description for the relevant properties should be extended to
mention, that not all cameras support all those values and theactual
values should be checked manually.