Add feature file load capability
This allows the user to give the path to an xml file containing all
camera features they want to set. It will be loaded and all settings
applied BEFORE all other settings provided as property are applied. This
should be included in the README to prevent confusion on the order of
feature changes. This is especially important for the `Width`, `Height`,
`OffsetX`, and `OffsetY` features as their default values will change
whatever is loaded from the XML file before.

This commit also includes a small change in how the camera ID is stored
internally. Before there was an additional field in
`vimbasrc->properties` which was used to store the default value of the
cameraid property. This was added by mistake while adding the structure
to hold ALL property values. Also the datatype for the
`vimbasrc->camera.id` field was adjusted to the more general `char`.