Do not overwrite features set from XML file
This commit changes the way feature values are written to the camera.
Instead of taking the settings from the XML file and potentially
overwriting them with values given for properties (or even the default
values for properties), no more features will be applied if an XML file
is passed. This means the user has to add all desired settings they want
to set on the camera to the XML file making it a single source of truth.