Fix infinite wait in `_create` function
The comparison to the return value seems to not always be correct here.
Just checking against the state is enough to see if the element should
continue to produce buffers or not.

This fixes a problem where the element would wait infinitely for a new
camera frame even if the pipeline requested a state change away from
`PLAYING`. This happens for example if camera is set up for hardware
triggering, but no trigger signals arrive or if the camera communication
is faulty and only incomplete frames arrive.