Add `vmb_open_count` to ensure corret VmbShutdown
Without this counting the `VmbShutdown()` call in
`gst_vimbasrc_finalize` is performed when the the first instance of the
`vimbasrc` element is closed. It should however be performed when the
last instance is finalized in order to allow communication to still
opened cameras. The `G_LOCK` and `G_UNLOCK` might be unnecessary but
make sure that this works even if GStreamer performs unexpected
switching in multithreading situations.