Search and replace `vimbasrc` -> `vimbaxsrc`
This was done in two steps to keep capitalization intact. This very
likely missed many occurences of the name vimba in the code, but it
makes the element useable for pipelines and sets the element name to
`vimbaxsrc` as expected.

The README etc. will have to be adjusted manually to ensure that context
is taken into account for every change.