Make path to xml available as `VmbFilePathChar_t`
This makes the call to `VmbSettingsLoad` succeed.

With the default settings decided by `VmbSettingsLoad` (i.e. not passing
a `VmbFeaturePersistSettings_t` pointer) I got an error that the max
iterations were reached. This implementation sets only those features
that are marked as `Streamable`, which is also the default behaviour in
e.g. VmbPy.