Add `allocationmode` property
This makes the element either use the "classic" Announce Mode (default)
where the element allocates the buffers and announces the frames to the
API, or the Alloc and Announce mode. In this second mode the buffer
pointers of the announced frames are `nullptr`s and the transport layer
is responsible for allocating and freeing the memory for the frame
transfers as needed.

The names for the modes were chosen to be similar to the names as
defined in `VmbCPP` where these modes were initially introduced.

Fixes OCT-3409