Round desired `OffsetX` and `OffsetY` to valid values
The simple calculation of `vmb_offsetx` and `vmb_offsety` can result in
values the camera might not support. By checking the `min`, `max` and
`increment` values for the correspoding features, this now rounds to the
nearest valid value for the features. This is deemed acceptable because
if the user lets the element automatically determine the offsets to
center the ROI, they are probably not interested if the actual set
offset ends up being 1 more or less than the exact center.